Preparing for Postgres procedures / functions.
-- procedure.rb, procedure_helpers.rb
- Moved the ProcedureClassMethods module to procedure_helpers.rb and
  renamed the module to ProcedureHelper.
- Also, moved the majority of the functionality that was contained in
  the ProcedureCommon module to procedure_helpers.rb and placed the
  methods in the module: ORAProcedureHelper.
- For the Procedure class a call is made to the procedure_helper method
  that got mixed in via ProcedureHelperProvider to provide (based on the
  dialect) the correct procedure methods (contained in either
  ORAProcedureHelper or PGProcedureHelper).

-- schema.rb, schema_helpers.rb, sequence.rb, sequence_helpers.rb
- Cleaned the HelperProvider classes a bit by only expecting a dialect
  to be passed as parameter. This way, most of the code has the same
  form when it comes to returning the helper classes to be mixed in.

-- procedure_spec.rb -> ora_procedure_spec.rb
- Renamed the procedure_spec.rb file to ora_procedure_spec.rb to
  indicate that the file tests Oracle functionality.

-- pg_procedure_spec.rb
- Introduced a pg_procedure_spec.rb file that will contain all the
  Postgres-specific procedure tests.
